requirements-dev.txt
black==24.10.0
flake8==7.1.0
pylint==3.3.3
mypy==1.14.0
isort==5.13.2
pre-commit==4.0.1
pycairo==1.27.0
